throw no network exception in business services

+++ Quick.Order.AppCore/BusinessOperations/BackOfficeRestaurantService.cs
@@ -1,4 +1,5 @@
 ﻿using Quick.Order.AppCore.Authentication.Contracts;
+using Quick.Order.AppCore.Contracts;
 using Quick.Order.AppCore.Contracts.Repositories;
 using Quick.Order.AppCore.Exceptions;
 using Quick.Order.AppCore.Models;
@@ -13,18 +14,26 @@ namespace Quick.Order.AppCore.BusinessOperations
         private readonly IRestaurantRepository restaurantRepository;
         private readonly IOrdersRepository ordersRepository;
         private readonly BackOfficeSessionService backOfficeSessionService;
+        private readonly IConnectivityService connectivityService;
 
-        public BackOfficeRestaurantService(IRestaurantRepository restaurantRepository,  IOrdersRepository ordersRepository, BackOfficeSessionService backOfficeSessionService)
+        public BackOfficeRestaurantService(IRestaurantRepository restaurantRepository,  IOrdersRepository ordersRepository, BackOfficeSessionService backOfficeSessionService, IConnectivityService connectivityService)
         {
             this.restaurantRepository = restaurantRepository;
             this.ordersRepository = ordersRepository;
             this.backOfficeSessionService = backOfficeSessionService;
+            this.connectivityService = connectivityService;
         }
         public async Task<Restaurant> AddRestaurant(Restaurant restaurant)
         {
             if (restaurant.Administrator == null)
             {
                 throw new SettingAdminForNewRestaurantException();
+            }
+
+            if (!connectivityService.HasNetwork())
+            {
+                throw new NoNetworkException();
+
             }
 
             var createdRestaurant= await restaurantRepository.Add(restaurant);
@@ -35,6 +44,11 @@ namespace Quick.Order.AppCore.BusinessOperations
 
         public async Task<bool> DeleteRestaurant(Restaurant restaurant)
         {
+            if (!connectivityService.HasNetwork())
+            {
+                throw new NoNetworkException();
+
+            }
             var result= await restaurantRepository.Delete(restaurant);
             backOfficeSessionService.CurrentRestaurantSession = null;
             return result;
@@ -42,19 +56,33 @@ namespace Quick.Order.AppCore.BusinessOperations
 
         public Task<bool> UpdateRestaurant(Restaurant restaurant)
         {
+            if (!connectivityService.HasNetwork())
+            {
+                throw new NoNetworkException();
+
+            }
             return restaurantRepository.Update(restaurant);
         }
 
        
         public Task<List<Models.Order>> GetOrdersForRestaurant(Guid restaurantId)
         {
-            
+            if (!connectivityService.HasNetwork())
+            {
+                throw new NoNetworkException();
+
+            }
             return ordersRepository.GetOrdersForRestaurant(restaurantId);
         }
 
 
         public Task SetOrderStatus(Models.Order order, OrderStatus status)
         {
+            if (!connectivityService.HasNetwork())
+            {
+                throw new NoNetworkException();
+
+            }
             order.OrderStatus = status;
             return ordersRepository.Update(order);
 
@@ -63,6 +91,11 @@ namespace Quick.Order.AppCore.BusinessOperations
 
         public Task<Models.Order> GetOrder(Guid id)
         {
+            if (!connectivityService.HasNetwork())
+            {
+                throw new NoNetworkException();
+
+            }
             return ordersRepository.GetById(id);
 
         }
@@ -70,6 +103,11 @@ namespace Quick.Order.AppCore.BusinessOperations
 
         public Task<bool> DeleteOrder(Models.Order order)
         {
+            if (!connectivityService.HasNetwork())
+            {
+                throw new NoNetworkException();
+
+            }
             return ordersRepository.Delete(order);
 
         }
